requirements.txt
-i https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/
anyio==3.7.1; python_version >= '3.7'
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
asttokens==2.4.1
attrs==23.1.0; python_version >= '3.7'
azure-ai-textanalytics==5.3.0
babel==2.13.1; python_version >= '3.7'
backcall==0.2.0
beautifulsoup4==4.8.2
bleach==6.1.0; python_version >= '3.8'
boto3==1.24.0; python_version >= '3.7'
botocore==1.27.96; python_version >= '3.7'
cachetools==5.3.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
comm==0.1.4; python_version >= '3.6'
contourpy==1.1.1; python_version >= '3.8'
cycler==0.12.1; python_version >= '3.8'
debugpy==1.8.0; python_version >= '3.8'
decorator==4.0.2; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
entrypoints==0.4; python_version >= '3.6'
et-xmlfile==1.1.0; python_version >= '3.6'
exceptiongroup==1.1.3; python_version < '3.11'
executing==2.0.1; python_version >= '3.5'
fastjsonschema==2.18.1
filelock==3.13.1; python_version >= '3.8'
fonttools==4.44.0; python_version >= '3.8'
fsspec==2023.6.0; python_version >= '3.8'
gdown==4.7.1
gensim==4.2.0; python_version >= '3.6'
google-api-core[grpc]==2.12.0; python_version >= '3.7'
google-auth==2.17.3; python_version >= '3.7'
google-cloud-language==2.4.2; python_version >= '3.6'
googleapis-common-protos==1.61.0; python_version >= '3.7'
grpcio==1.59.2
grpcio-status==1.59.2
huggingface-hub==0.18.0; python_full_version >= '3.8.0'
idna==3.4; python_version >= '3.5'
importlib-metadata==6.8.0; python_version < '3.10'
importlib-resources==5.9.0; python_version >= '3.7'
ipympl==0.9.2
ipython-genutils==0.2.0
ipywidgets==7.7.2
isodate==0.6.1
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
joblib==1.3.2; python_version >= '3.7'
json5==0.9.14
jsonschema==4.19.2; python_version >= '3.8'
jsonschema-specifications==2023.7.1; python_version >= '3.8'
jupyter-client==7.4.9; python_version >= '3.7'
jupyter-core==5.5.0; python_version >= '3.8'
jupyter-server==1.24.0; python_version >= '3.7'
jupyterlab==3.4.3; python_version >= '3.7'
jupyterlab-pygments==0.2.2; python_version >= '3.7'
jupyterlab-server==2.25.0; python_version >= '3.8'
jupyterlab-widgets==1.1.7; python_version >= '3.6'
kaleido==0.2.1
kiwisolver==1.4.5; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
matplotlib==3.7.3; python_version >= '3.8'
matplotlib-inline==0.1.6; python_version >= '3.5'
mistune==3.0.2; python_version >= '3.7'
msrest==0.7.1; python_version >= '3.6'
nbclassic==0.5.6; python_version >= '3.7'
nbclient==0.8.0; python_full_version >= '3.8.0'
nbconvert==7.10.0; python_version >= '3.8'
nbformat==5.9.2; python_version >= '3.8'
nest-asyncio==1.5.8; python_version >= '3.5'
nlpaug==1.1.11; python_version >= '3.7'
nltk==3.7; python_version >= '3.7'
notebook==6.5.5; python_version >= '3.7'
notebook-shim==0.2.3; python_version >= '3.7'
numpy; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
openpyxl==3.0.10; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pandas==1.5.3; python_version >= '3.8'
pandocfilters==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==10.1.0; python_version >= '3.8'
pkgutil-resolve-name==1.3.10; python_version < '3.9'
platformdirs==3.11.0; python_version >= '3.7'
plotly==5.11.0; python_version >= '3.6'
prometheus-client==0.18.0; python_version >= '3.8'
proto-plus==1.22.3; python_version >= '3.6'
protobuf==4.25.0; python_version >= '3.8'
psutil==5.9.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pygments==2.16.1; python_version >= '3.7'
pyparsing==3.1.1; python_full_version >= '3.6.8'
pysocks==1.7.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.30.2; python_version >= '3.8'
regex==2023.10.3; python_version >= '3.7'
requests[socks]==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rpds-py==0.12.0; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
s3transfer==0.6.2; python_version >= '3.7'
scikit-learn==1.3.2; python_version >= '3.8'
scipy==1.10.1; python_version < '3.12' and python_version >= '3.8'
seaborn==0.12.1; python_version >= '3.7'
send2trash==1.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smart-open==6.4.0; python_version >= '3.6' and python_version < '4.0'
sniffio==1.3.0; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
stack-data==0.6.3
tenacity==8.2.3; python_version >= '3.7'
terminado==0.17.1; python_version >= '3.7'
threadpoolctl==3.2.0; python_version >= '3.8'
tinycss2==1.2.1; python_version >= '3.7'
tokenizers==0.12.1
tqdm==4.66.1; python_version >= '3.7'
traitlets==5.13.0; python_version >= '3.8'
transformers==4.20.1; python_full_version >= '3.7.0'
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==1.26.18; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
wcwidth==0.2.9
webencodings==0.5.1
websocket-client==1.6.4; python_version >= '3.8'
widgetsnbextension==3.6.6
xlrd==2.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
zipp==3.17.0; python_version < '3.10'
fastapi
python-multipart
uvicorn
cohere
openai
tiktoken